+++ ktlint-ruleset-experimental/src/main/kotlin/com/pinterest/ktlint/ruleset/experimental/ExperimentalRuleSetProvider.kt
@@ -33,6 +33,7 @@ public class ExperimentalRuleSetProvider : RuleSetProvider {
         KdocWrappingRule(),
         SpacingBetweenFunctionNameAndOpeningParenthesisRule(),
         ParameterListSpacingRule(),
-        FunctionReturnTypeSpacingRule()
+        FunctionReturnTypeSpacingRule(),
+        FunctionStartOfBodySpacingRule()
     )
 }